debugging ingest_transit
