+++ src/mjolnir/CMakeLists.txt
@@ -31,6 +31,7 @@ set(sources
   edgeinfobuilder.cc
   ferry_connections.cc
   graphfilter.cc
+  ingest_transit.cc
   linkclassification.cc
   node_expander.cc
   osmdata.cc
@@ -83,6 +84,7 @@ valhalla_module(NAME mjolnir
    PRIVATE
       ${CMAKE_CURRENT_BINARY_DIR}
       ${CMAKE_CURRENT_BINARY_DIR}/valhalla
+      ${VALHALLA_SOURCE_DIR}/third_party/just_gtfs/include
   DEPENDS
     valhalla::proto
     valhalla::baldr
